Remove redundant network subpackage
