+++ tests/integration_tests/network/test_http.py
@@ -1,4 +1,4 @@
-# -------------------------------------------------------------------------------------------------
+# -------------------------------------------------------------------------------------------------G
 #  Copyright (C) 2015-2023 Nautech Systems Pty Ltd. All rights reserved.
 #  https://nautechsystems.io
 #
@@ -13,60 +13,107 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
-import asyncio
 from collections.abc import Coroutine
+from typing import Any, Callable
 
+import msgspec
 import pytest
-import pytest_asyncio
 from aiohttp import web
 from aiohttp.test_utils import TestServer
 
-from nautilus_trader.network.http import HttpClient
-from nautilus_trader.test_kit.stubs.component import TestComponentStubs
+from nautilus_trader.core.nautilus_pyo3.network import HttpClient
+from nautilus_trader.core.nautilus_pyo3.network import HttpResponse
 
 
-@pytest.fixture(name="server")
-async def fixture_server(aiohttp_server):
+@pytest.fixture(name="test_server")
+async def fixture_test_server(
+    aiohttp_server: Callable[..., Coroutine[Any, Any, TestServer]],
+) -> TestServer:
     async def hello(request):
         return web.Response(text="Hello, world")
 
     app = web.Application()
     app.router.add_route("GET", "/get", hello)
     app.router.add_route("POST", "/post", hello)
+    app.router.add_route("PATCH", "/patch", hello)
+    app.router.add_route("DELETE", "/delete", hello)
 
     server = await aiohttp_server(app)
     return server
 
 
-@pytest_asyncio.fixture(name="client")
-async def fixture_client() -> HttpClient:
-    client = HttpClient(
-        loop=asyncio.get_event_loop(),
-        logger=TestComponentStubs.logger(),
-    )
-    await client.connect()
-    return client
+@pytest.mark.asyncio()
+async def test_client_get(test_server: Coroutine) -> None:
+    # Arrange
+    server: TestServer = await test_server
+    client = HttpClient()
+    url = f"http://{server.host}:{server.port}/get"
+
+    # Act
+    response: HttpResponse = await client.get(url, headers={})
+
+    # Assert
+    assert response.status == 200
+    assert len(response.body) > 0
 
 
 @pytest.mark.asyncio()
-async def test_client_get(client: HttpClient, server: Coroutine) -> None:
-    test_server: TestServer = await server
-    url = f"http://{test_server.host}:{test_server.port}/get"
-    resp = await client.get(url)
-    assert resp.status == 200
-    assert len(resp.data) > 0
-    assert client.max_latency() > 0
-    assert client.min_latency() > 0
-    assert client.avg_latency() > 0
+async def test_client_post(test_server: Coroutine) -> None:
+    # Arrange
+    server: TestServer = await test_server
+    client = HttpClient()
+    url = f"http://{server.host}:{server.port}/post"
+
+    # Act
+    response: HttpResponse = await client.post(url, headers={})
+
+    # Assert
+    assert response.status == 200
+    assert len(response.body) > 0
 
 
 @pytest.mark.asyncio()
-async def test_client_post(client: HttpClient, server: Coroutine) -> None:
-    test_server = await server
-    url = f"http://{test_server.host}:{test_server.port}/post"
-    resp = await client.post(url)
-    assert resp.status == 200
-    assert len(resp.data) > 10
-    assert client.max_latency() > 0
-    assert client.min_latency() > 0
-    assert client.avg_latency() > 0
+async def test_client_post_with_body(test_server: Coroutine) -> None:
+    # Arrange
+    server: TestServer = await test_server
+    client = HttpClient()
+    url = f"http://{server.host}:{server.port}/post"
+    body = {"key1": "value1", "key2": "value2"}
+    body_bytes = msgspec.json.encode(body)
+
+    # Act
+    response: HttpResponse = await client.post(url, headers={}, body=body_bytes)
+
+    # Assert
+    assert response.status == 200
+    assert len(response.body) > 0
+
+
+@pytest.mark.asyncio()
+async def test_client_patch(test_server: Coroutine) -> None:
+    # Arrange
+    server: TestServer = await test_server
+    client = HttpClient()
+    url = f"http://{server.host}:{server.port}/patch"
+
+    # Act
+    response: HttpResponse = await client.patch(url, headers={})
+
+    # Assert
+    assert response.status == 200
+    assert len(response.body) > 0
+
+
+@pytest.mark.asyncio()
+async def test_client_delete(test_server: Coroutine) -> None:
+    # Arrange
+    server: TestServer = await test_server
+    client = HttpClient()
+    url = f"http://{server.host}:{server.port}/delete"
+
+    # Act
+    response: HttpResponse = await client.delete(url, headers={})
+
+    # Assert
+    assert response.status == 200
+    assert len(response.body) > 0
